% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bbox_coords}
\alias{bbox_coords}
\title{A function to create bbox objects to feed into planet_search()}
\usage{
bbox_coords(coords, outputtype = "sf", bufferdist = 0.01)
}
\arguments{
\item{coords}{a vector of longitude latitude values, format: c(long1, long2, lat1, lat2) or c(long1,lat1)}

\item{outputtype}{a character string specifying sf::st_bbox (DEFAULT) or terra::ext  as the output format}

\item{bufferdist}{a distance in unit lat/long degrees for buffering if only a single point is provided in `coords`. Defaults to 0.01 degrees approximately 1 kilometre.}
}
\description{
This function can take input values and create a bbox to be fed to the `planet_search()` function.
}
\keyword{Planet}
