% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet_activate.R
\name{planet_activate}
\alias{planet_activate}
\title{A function to batch activate the results from planet_search()}
\usage{
planet_activate(i, item_name = "PSOrthoTile")
}
\arguments{
\item{i}{Row index (integer) of the search result. Mant to be used in a loop.}

\item{item_name}{Defaults to "PSOrthoTile".}
}
\description{
This function allows you to activate assets using the Planet API. Assets cannot be downloaded until activated.
}
\keyword{Planet}
