% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet_download.R
\name{planet_download}
\alias{planet_download}
\title{A function to download activated Planet imagery}
\usage{
planet_download(i, overwrite = T)
}
\arguments{
\item{bbox}{shapefile of bounding box must be EPSG:4326 Projection; no default.}

\item{date_end}{Expects as.Date; defaults to as.Date('2018-07-01')}

\item{date_start}{Expects as.Date; defaults to as.Date('2018-08-01')}

\item{cloud_lim}{Cloud percentage from 0-1; defaults to 0.1, or 10 percent.}

\item{item_name}{Defaults to "PSOrthoTile".}
}
\description{
This function allows you to search the Planet API
}
\keyword{Planet}
