% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet_order.R
\name{planet_order}
\alias{planet_order}
\title{A function to order Planet imagery}
\usage{
planet_order(
  api_key,
  bbox,
  date_start,
  date_end,
  list_dates = NULL,
  cloud_lim,
  item_name = "PSScene",
  product_bundle = "analytic_sr_udm2",
  asset = "ortho_analytic_4b_sr",
  order_name,
  mostrecent = 0
)
}
\arguments{
\item{api_key}{a string containing your API Key for your planet account}

\item{bbox}{bounding box made with extent() from the raster package; must be EPSG:4326 Projection; no default.}

\item{date_start}{a date object}

\item{date_end}{doy start}

\item{cloud_lim}{Cloud percentage from 0-1; defaults to 0.1, or 10 percent.}

\item{item_name}{Defaults to "PSScene".}

\item{product_bundle}{Defaults to "analytic_sr_udm2}

\item{asset}{Defaults to "ortho_analytic_4b_sr"}

\item{order_name}{The name you want to assign to your order}

\item{mostrecent}{Integer of how many of the most recent images will be downloaded. Default is 0 (download all images).}
}
\description{
This function allows you to download orders from the Planet Orders API
}
\keyword{Planet}
