% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet_order.R
\name{planet_order_download}
\alias{planet_order_download}
\title{A function to order Planet imagery}
\usage{
planet_order_download(order_id, exportfolder, api_key)
}
\arguments{
\item{order_id}{request order id (output from `planet_order_request()`)}

\item{exportfolder}{The name you want to assign to your order}
}
\description{
This function allows you to download orders from the Planet Orders API
}
\keyword{Planet}
