% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet_order.R
\name{planet_order_request}
\alias{planet_order_request}
\title{A function to order Planet imagery}
\usage{
planet_order_request(
  api_key,
  bbox,
  date_start,
  date_end,
  list_dates = NULL,
  cloud_lim,
  item_name,
  product_bundle,
  asset,
  order_name = exportfolder,
  mostrecent
)
}
\arguments{
\item{api_key}{a string containing your API Key for your planet account}

\item{bbox}{bounding box made with extent() from the raster package; must be EPSG:4326 Projection; no default.}

\item{date_start}{a date object}

\item{date_end}{a date object}

\item{cloud_lim}{Cloud percentage from 0-1; defaults to 0.1, or 10 percent.}

\item{item_name}{Defaults to "PSScene4Band".}

\item{product_bundle}{Defaults to "analytic_sr"}

\item{order}{The name you want to assign to your order. Defaults to "AutomationTEST"}
}
\description{
This function allows you to search and activate orders from the Planet Orders API
}
\keyword{Planet}
