% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{planet_order_request_items}
\alias{planet_order_request_items}
\title{A function to order Planet imagery}
\usage{
planet_order_request_items(
  items,
  bbox,
  item_name = "PSScene4Band",
  product_bundle = "analytic_sr",
  asset = "ortho_analytic_4b_sr",
  order_name,
  api_key
)
}
\arguments{
\item{items}{a vector containing items to request}

\item{bbox}{bounding box made with ext() from the terra package; must be EPSG:4326 Projection; no default.}

\item{item_name}{Defaults to "PSScene4Band".}

\item{product_bundle}{Defaults to "analytic_sr"}

\item{asset}{Defaults to "ortho_analytic_4b_sr"}

\item{order_name}{The name you want to assign to your order.}

\item{api_key}{a string containing your API Key for your planet account}
}
\description{
This function allows you to search and activate orders from the Planet Orders API
}
\keyword{Planet}
