% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet_search.R
\name{planet_search}
\alias{planet_search}
\title{A function to search Planet imagery}
\usage{
planet_search(
  bbox = bbox,
  date_end = NULL,
  date_start = NULL,
  cloud_lim = 0.1,
  ground_control = TRUE,
  quality = "standard",
  item_name = "PSScene",
  asset = "ortho_analytic_8b_sr",
  api_key = "test",
  list_dates = NULL
)
}
\arguments{
\item{bbox}{bounding box made with extent() from the raster package; must be EPSG:4326 Projection; no default.}

\item{date_end}{Expects as.Date; defaults to as.Date('2018-07-01')}

\item{date_start}{Expects as.Date; defaults to as.Date('2018-08-01')}

\item{cloud_lim}{Cloud percentage from 0-1; defaults to 0.1, or 10 percent.}

\item{ground_control}{Defaults to TRUE, filter images to only those with ground control, ensures locational accuracy of 10 m RMSE or better}

\item{quality}{Defaults to "standard", other option is "test" see https://support.planet.com/hc/en-us/articles/4407808871697-Image-quality-Standard-vs-Test-imagery}

\item{item_name}{Defaults to "PSOrthoTile".}

\item{api_key}{your planet api key string}
}
\description{
This function allows you to search the Planet API
}
\keyword{Planet}
