% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{planet_activate}
\alias{planet_activate}
\title{A function to batch activate the results from planet_search()}
\usage{
planet_activate(
  item,
  item_name = "PSOrthoTile",
  asset = "ortho_analytic_4b_sr",
  api_key
)
}
\arguments{
\item{item}{ID of one item (output from `planet_search()`)}

\item{item_name}{Defaults to "PSOrthoTile".}

\item{asset}{Defaults to "ortho_analytic_4b_sr"}

\item{api_key}{a string containing your API Key for your planet account}
}
\description{
This function allows you to activate assets using the Planet 'Data' API. Assets cannot be downloaded until activated. Function meant to be performed in a loop.
}
\keyword{Planet}
