% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{planet_download}
\alias{planet_download}
\title{A function to download activated Planet imagery}
\usage{
planet_download(
  item,
  item_name = "PSOrthoTile",
  asset = "ortho_analytic_8b_sr",
  exportfolder,
  api_key,
  overwrite = T
)
}
\arguments{
\item{item}{ID of one item (output from `planet_search()`)}

\item{item_name}{Defaults to "PSOrthoTile".}

\item{asset}{Defaults to "ortho_analytic_8b_sr"}

\item{exportfolder}{The name of the directory you want your files to be downloaded into}

\item{api_key}{a string containing your API Key for your planet account}

\item{overwrite}{Defaults to TRUE. Overwrite file in `exportfolder`?}
}
\description{
This function allows you to download imagery using the Planet 'Data' API, after use of the `planet_activate()` function.
}
\keyword{Planet}
