% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{planet_order_request}
\alias{planet_order_request}
\title{A function to order Planet imagery}
\usage{
planet_order_request(
  bbox,
  date_start,
  date_end,
  list_dates = NULL,
  cloud_lim = 0.1,
  item_name = "PSScene4Band",
  product_bundle = "analytic_sr",
  asset = "ortho_analytic_4b_sr",
  order_name,
  mostrecent,
  api_key
)
}
\arguments{
\item{bbox}{bounding box made with ext() from the terra package; must be EPSG:4326 Projection; no default.}

\item{date_start}{a date object}

\item{date_end}{a date object}

\item{list_dates}{Default to NULL. a vector of dates, to substitute `date_start` and `date_end`}

\item{cloud_lim}{Cloud percentage from 0-1; defaults to 0.1, or 10 percent.}

\item{item_name}{Defaults to "PSScene4Band".}

\item{product_bundle}{Defaults to "analytic_sr"}

\item{asset}{Defaults to "ortho_analytic_4b_sr"}

\item{order_name}{The name you want to assign to your order. Defaults to "AutomationTEST"}

\item{mostrecent}{Integer of how many of the most recent images will be downloaded. Default is 0 (download all images).}

\item{api_key}{a string containing your API Key for your planet account}
}
\description{
This function allows you to search and activate orders from the Planet 'Orders' API. This function wraps the `planet_search()` function and places an order for the specified number of 'most recent' images.
}
\keyword{Planet}
