% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{planet_order_request_items}
\alias{planet_order_request_items}
\title{A function to order Planet imagery}
\usage{
planet_order_request_items(
  items,
  clip = TRUE,
  bbox = NULL,
  item_name = "PSScene4Band",
  product_bundle = "analytic_sr",
  asset = "ortho_analytic_8b_sr",
  order_name,
  api_key
)
}
\arguments{
\item{items}{a vector containing items to request: an output of `planet_search()`}

\item{clip}{Indicates whether to CLIP *each* image in the order to the *same* bbox. Defaults to TRUE. Should only be used if images cover the same geographic extent.}

\item{bbox}{Used if clip=TRUE. Bounding box made with ext() from the terra package; must be EPSG:4326 Projection; no default.}

\item{item_name}{Defaults to "PSScene".}

\item{product_bundle}{Defaults to "analytic_sr_udm2}

\item{asset}{Defaults to "ortho_analytic_8b_sr"}

\item{order_name}{The name you want to assign to your order.}

\item{api_key}{a string containing your API Key for your planet account}
}
\description{
This function allows you to search and activate orders from the Planet 'Orders' API. This function allows for manual selection of items from the `planet_search()` function
}
\keyword{Planet}
